# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Sulocraft API
  description: Backend service for the Sulocraft direct-to-consumer crochet storefront.
  version: 0.2.0
paths:
...
  /api/v1/auth/email/start:
    post:
      tags:
      - auth
      summary: Start Email Login
      description: Send a single-use magic sign-in link via email. Always returns
        generic non-enumerating 202.
      operationId: start_email_login_api_v1_auth_email_start_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailStartRequest'
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailStartResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/email/verify:
    get:
      tags:
      - auth
      summary: Verify Email Magic Link
      description: Consume a magic-link token, authenticate/link user, merge guest
        cart, and redirect.
      operationId: verify_email_magic_link_api_v1_auth_email_verify_get
      parameters:
      - name: token
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
      - name: returnTo
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Returnto
      - name: x-cart-token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Cart-Token
      - name: guest_cart_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Guest Cart Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/google:
    post:
      tags:
      - auth
      summary: Google Auth
      description: One-Click Google Authentication with Unified User Model.
      operationId: google_auth_api_v1_auth_google_post
      parameters:
      - name: x-cart-token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Cart-Token
      - name: guest_cart_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Guest Cart Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleAuthRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/facebook:
    post:
      tags:
      - auth
      summary: Facebook Auth
      description: Facebook (Meta) Authentication with Unified User Model.
      operationId: facebook_auth_api_v1_auth_facebook_post
      parameters:
      - name: x-cart-token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Cart-Token
      - name: guest_cart_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Guest Cart Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FacebookAuthRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/me:
    get:
      tags:
      - auth
      summary: Get Current Customer
      description: Retrieve authenticated customer profile from session cookie.
      operationId: get_current_customer_api_v1_auth_me_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: x-session-token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Session-Token
      - name: session_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/logout:
    post:
      tags:
      - auth
      summary: Logout
      description: Revoke session and clear HttpOnly cookie.
      operationId: logout_api_v1_auth_logout_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: x-session-token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Session-Token
      - name: session_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                title: Response Logout Api V1 Auth Logout Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/orders:
    post:
      tags:
      - orders
      summary: Checkout
      description: 'Convert the current shopping cart into a placed order.


        Snapshots products, variant details, prices, and shipping address.

        Decrements inventory, clears active cart items, and logs initial status history.'
      operationId: checkout_api_v1_orders_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: x-session-token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Session-Token
      - name: x-cart-token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Cart-Token
      - name: session_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Token
      - name: guest_cart_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Guest Cart Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - orders
      summary: List Orders
      description: Retrieve order history for the authenticated customer.
      operationId: list_orders_api_v1_orders_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: x-session-token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Session-Token
      - name: session_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderOut'
                title: Response List Orders Api V1 Orders Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AuthResponse:
      properties:
        user:
          $ref: '#/components/schemas/UserOut'
        message:
          type: string
          title: Message
          default: Authentication successful
        cartMerged:
          type: boolean
          title: Cartmerged
          default: false
      type: object
      required:
      - user
      title: AuthResponse
      description: Authentication success response.
...
    EmailStartRequest:
      properties:
        email:
          type: string
          title: Email
          description: Customer email address.
      type: object
      required:
      - email
      title: EmailStartRequest
      description: Payload to request an email magic link.
    EmailStartResponse:
      properties:
        message:
          type: string
          title: Message
          default: If the email address is valid, a sign-in link has been sent.
        devMagicLink:
          anyOf:
          - type: string
          - type: 'null'
          title: Devmagiclink
          description: Provided in local/test mode only.
      type: object
      title: EmailStartResponse
      description: Generic non-enumerating response for magic link initiation.
...
    GoogleAuthRequest:
      properties:
        credential:
          type: string
          title: Credential
          description: Google ID token credential from Google Sign-In button.
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
          description: User full name.
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
          description: User email address.
        sub:
          anyOf:
          - type: string
          - type: 'null'
          title: Sub
          description: Google Subject / user ID.
      type: object
      required:
      - credential
      title: GoogleAuthRequest
      description: Payload for One-Click Google Authentication.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...